Added SSH key on security key
This SSH key is used to connect to my machines from a less trusted client.
It is important to note that this is not to be used on completely untrusted clients, rather just *less* trusted clients.
The use is not for security through less client trust, but rather security through key impermanence, preventing a client being compromised after a session from being an issue.